requirements.txt
python-telegram-bot==22.5
requests==2.32.5
python-dotenv==1.0.0
APScheduler==3.10.4
flask==3.0.0